@@ -1,10 +1,10 @@
 package mvvm.whoami.mvvmfast.base
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v4.app.FragmentActivity
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.FragmentActivity
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProviders
 import com.noober.background.BackgroundLibrary
 import me.jessyan.autosize.AutoSizeConfig
 import mvvm.whoami.mvvmfast.widget.LoadingView
